Move containers API, fix endpoint API for FAIR
