+++ tests/routes/health_test.py
@@ -1,6 +1,6 @@
 import mock
 import pytest
-from time import sleep
+
 
 from flask import Flask, appcontext_pushed, g
 
@@ -12,7 +12,7 @@ from web.models.schain import SChainRecord
 from web.routes.health import health_bp
 from web.helper import get_api_url
 
-from tests.utils import get_bp_data, get_schain_struct, run_custom_schain_container
+from tests.utils import get_bp_data, get_schain_struct
 
 
 TEST_SGX_KEYNAME = 'test_keyname'
@@ -55,34 +55,6 @@ def unregistered_skale_bp(skale, dutils):
             yield app.test_client()
         finally:
             SChainRecord.drop_table()
-
-
-def test_containers(skale_bp, dutils):
-    data = get_bp_data(skale_bp, get_api_url('health', 'containers'))
-    expected = {
-        'status': 'ok',
-        'payload': dutils.get_containers_info(all=False, name_filter='', format=True),
-    }
-    assert data == expected
-    for container_info in data['payload']:
-        field_map = {'cpu_shares': 0, 'mem_limit': 0, 'swap_limit': 0, 'swappiness': None}
-        for field in field_map.keys():
-            assert field in container_info
-
-
-def test_containers_all(skale_bp, dutils, schain_db, cleanup_schain_containers):
-    run_custom_schain_container(dutils, schain_db, 'bash -c "exit 1"')
-    sleep(10)
-    data = get_bp_data(skale_bp, get_api_url('health', 'containers'), params={'all': True})
-    expected = {
-        'status': 'ok',
-        'payload': dutils.get_containers_info(all=True, name_filter='', format=True),
-    }
-    assert data == expected
-    for container_info in data['payload']:
-        field_map = {'cpu_shares': 0, 'mem_limit': 0, 'swap_limit': 0, 'swappiness': None}
-        for field in field_map.keys():
-            assert field in container_info
 
 
 def test_schains_checks(skale_bp, skale, schain_on_contracts, schain_db, dutils):
